fix: 금액 long으로 타입 변경
- int일 경우 범위 초과 발생
- Result에 LottoCompareResult를 추가하는 작업 getTotalWinningMoney 메소드 내부로 이동

+++ src/main/java/lotto/Rank.java
@@ -12,9 +12,9 @@ public enum Rank {
 
     private final int countOfMatch;
     private final boolean matchBonus;
-    private final int winningMoney;
+    private final long winningMoney;
 
-    Rank(int countOfMatch, boolean matchBonus, int winningMoney) {
+    Rank(int countOfMatch, boolean matchBonus, long winningMoney) {
         this.countOfMatch = countOfMatch;
         this.matchBonus = matchBonus;
         this.winningMoney = winningMoney;
@@ -24,7 +24,7 @@ public enum Rank {
         return countOfMatch;
     }
 
-    public int getWinningMoney() {
+    public long getWinningMoney() {
         return winningMoney;
     }
